backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
openai==1.55.3
httpx==0.27.2
beautifulsoup4==4.12.3
langdetect==1.0.9
faiss-cpu==1.9.0
numpy==2.1.3
langchain==0.3.9
langchain-openai==0.2.6
supabase==2.10.0
python-dotenv==1.0.1
watchdog==6.0.0
